% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation.R
\name{assignStates}
\alias{assignStates}
\title{assignStates()}
\usage{
assignStates(object)
}
\arguments{
\item{object}{casper object}
}
\value{
object
}
\description{
calculates baf shift threshold using gaussian mixture models and assigns deletion or amplification to a segment when the HMM state is 1 or 5 without looking at the BAF signal. When the segment state is 2 or 4, an accompanying BAF shift on the segment is required.
}
