% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{AverageReference}
\alias{AverageReference}
\title{AverageReference()}
\usage{
AverageReference(data, ref_ids)
}
\arguments{
\item{object}{casper object}
}
\value{
object
}
\description{
the mean the expression level for each gene across all the reference cells (samples) are computed.
}
