\name{calcDenovo}
\alias{calcDenovo}
\alias{relativeExpr}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
  Estimate expression of gene splicing variants de novo. 
}
\description{
  \code{calcDenovo} estimates expression of gene splicing variants,
  considering both known variants and variants that have not been
  previously described.

  \code{relativeExpr} computes relative expression estimates from
  \code{calcDenovo} output via Bayesian model selection or averaging.
}
\usage{
calcDenovo(distrs, genomeDB, pc, readLength, islandid, priorq=3, mprior,
minpp=0.001, selectBest=FALSE, method="auto", niter, exactMarginal=TRUE,
verbose=TRUE, mc.cores=1)

relativeExpr(expr, summarize="modelAvg", minProbExpr=0.5, minExpr=0.05)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
\item{distrs}{ List of fragment distributions as generated by the \code{getDistrs} function}
\item{genomeDB}{ \code{denovoGenome} object containing annotated genome,
  as returned by the \code{createDenovoGenome} function. }
\item{pc}{Named vector of exon path counts as returned by \code{pathCounts}}
\item{readLength}{ Read length in bp, e.g. in a paired-end experiment where
  75bp are sequenced on each end one would set \code{readLength=75}.}
\item{islandid}{Name of the gene island to be analyzed. If not specified, all
  gene islands are analyzed.}
\item{priorq}{Parameter of the prior distribution on the proportion of reads coming from each variant. The prior is Dirichlet with prior sample size for each variant equal to priorq.
            We recommend \code{priorq=3} as this defines a non-local
	    prior that penalizes including variants with low estimated
	    expression unless they are warranted by the data.}
\item{mprior}{Prior on the model space, as returned by
  \code{modelPrior}. This argument is used to favor splicing variant
  configurations that are consistent with the annotated genome. If left
  missing a uniform prior on the model space is used, i.e. all splicing
  variants configurations are equally likely a priori. We strongly
  recommend setting \code{mprior} via \code{modelPrior}. See details.}
\item{minpp}{Models (i.e. splicing configurations) with posterior probability less than \code{minpp}
  are not reported. This argument can help reduce substantially the amount of
  required memory to store the results.}
\item{selectBest}{If set to \code{TRUE} only the model with highest
  posterior probability is reported. While this can save memory, we do
  not recommend this option as it may ignore a substantial amount of
  uncertainty.}
\item{method}{Method used to perform the model search. 
  \code{"exact"} exhaustively enumerates all possible models (warning:
  this is not feasible for genes over 5 exons). \code{"rwmcmc"} uses a
  random-walk MCMC scheme to focus on models with high posterior
  probability. \code{"auto"} uses \code{"exact"} for genes with up to 5
  exons and \code{"rwmcmc"} for longer genes. See details.}
\item{niter}{Number of MCMC iterations.}
\item{exactMarginal}{Set to \code{FALSE} to estimate posterior model
  probabilities as the proportion of MCMC visits. Set to \code{TRUE} to
  use the integrated likelihoods (default). See details.}
\item{verbose}{Set to \code{TRUE} to display progress information.}
\item{mc.cores}{Number of processors to be used for parallel
  computation. Can only be used if the package \code{multicore} is
  available for your system.
Warning: using multiple processors substantially increases the memory
requirements, so set this value carefully.}

\item{expr}{Object of class denovoGenomeExpr, as returned by \code{calcDenovo}.}
\item{summarize}{Set to \code{"modelAvg"} to average estimates across models
according to the posterior probability of each model. Set to
\code{"bestModel"} to report the posterior mode, i.e. the estimate under
the model with highest posterior probabiliby.}
\item{minProbExpr}{Variants with marginal posterior probability of being
expressed below \code{minProbExpr} are not reported.}
\item{minExpr}{Variants with estimated expression below \code{minExpr}
are not reported.}
}
\details{
  \code{calcDenovo} attempts to consider all possible splicing variant
  configurations for each gene. For example, a gene with 5 exons has 31
  possible variants (2^5 - 1), each of which may be either expressed or
  not expressed. This gives rise to 2^31 - 1 possible configurations of
  expressed variants.

  The routines computes the posterior probability of each model
  (i.e. configuration of expressed variants) via Bayes theorem.

  P(model|y) "proportional to" m(y|model) P(model)

  where m(y|model) is the integrated likelihood and P(model) is the
  prior probability of the model. Importantly, P(model) can be set by analyzing available genome
  annotations. For instance, genes with 20 exons have variants that tend
  to have close to 20 exons. They also tend to have more expressed
  variants than genes with 5 exons. The function \code{modelPrior}
  analyzes an annotated genome to set reasonable values for P(model).

  Given that the number of models grows super-exponentially with the
  number of exons, an exhaustive enumeration of all possible models is
  not feasible unless the gene is very short (in practice, we set the
  limit at 5 exons). For longer genes, we use a random walk MCMC scheme
  that incorporates Shotgun Stochastic Search. This random exploration
  scheme focuses on models with large posterior probability. The longer
  the number of iterations (\code{niter}) the more stable the results,
  but this can come at a substantial computational cost.

  In order to compute P(model|y) one can either use the computed
  m(y|model) P(model) (option \code{exactMarginal==TRUE}) or the
  proportion of MCMC visits (option \code{exactMarginal==FALSE}). Unless
  \code{niter} is large the former option typically provides more
  precise estimates.
}
\value{
  Expression set with expression estimates.
  \code{featureNames} identify each transcript via
  RefSeq ids, and the \code{featureData} contains Entrez identifiers
  and the length of each transcript (in bp).
  If \code{citype} was set to a value other than \code{"none"}, the \code{featureData} also contains the 95\% credibility intervals
  (i.e. intervals that contain the true parameter value with 95\% posterior probability).
}
\references{
Rossell D. QASPER: Quantifying Alternative Splicing from Paired End
Reads. Technical report, IRB Barcelona (2010)
}
\author{
  Camille Stephan-Otto Attolini, Manuel Kroiss, David Rossell
}
\seealso{
\code{plotExpr} to produce a plot with splicing variants and estimated expression.
}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ models }% __ONLY ONE__ keyword per line
