% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{calcROC}
\alias{calcROC}
\title{calcROC()}
\usage{
calcROC(chrMat, chrMat2)
}
\arguments{
\item{chrMat}{large scale event matrix generated using CaSpER}

\item{chrMat2}{large scale event matrix generated using genotyping array}
}
\value{
accuracy measures
}
\description{
Calculates tpr and fpr values using genotyping array as gold standard
}
