% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation.R
\name{calculateLOHShiftsForEachSegment}
\alias{calculateLOHShiftsForEachSegment}
\title{calculateLOHShiftsForEachSegment()}
\usage{
calculateLOHShiftsForEachSegment(object)
}
\arguments{
\item{object}{casper object}
}
\value{
object
}
\description{
calculate the median value of the BAF shift signal on the segments
}
