% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casper.R
\docType{package}
\name{CaSpER-package}
\alias{CaSpER-package}
\alias{_PACKAGE}
\title{CaSpER: Identify large-scale CNV events from single cell or bulk RNA-Seq data}
\description{
Identification, visualization and integrative analysis of CNV events in multiscale resolution using single-cell or bulk RNA sequencing data
}
\details{
The main functions you will need to use are CreateCasperObject() and runCaSpER(casper_object).
For additional details on running the analysis step by step, please refer to the example vignette.
}
