% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casper.R
\docType{class}
\name{casper}
\alias{casper}
\alias{casper-class}
\title{The CaSpER Class}
\description{
The CaSpER Class
The casper object is required for performing CNV analysis on single-cell and bulk RNA-Seq. It stores all information
associated with the dataset, including data, smoothed data, baf values, annotations, scale specific segments, scale specific large scale events etc.
}
\section{Slots}{

\describe{
\item{\code{raw.data}}{raw project data}

\item{\code{data}}{lowly expressed genes are filtered from the data}

\item{\code{loh}}{original baf signal}

\item{\code{median.filtered.data}}{median filtered expression signal}

\item{\code{loh.median.filtered.data}}{median filtered baf signal}

\item{\code{centered.data}}{gene expression levels are centered around the mid-point. For each gene, the mid-point of expression level is computed among all the cells (or samples in bulk RNA-seq), then the mid-point expression level is subtracted from the expression levels}

\item{\code{center.smoothed.data}}{cell centric expression centering is performed. For each cell (or sample), we compute the mid-point of the expression level then we subtract the mid-point expression from the expression levels of all the genes for the corresponding cel}

\item{\code{control.normalized}}{control normalization is performed by subtracting reference expression values from the tumor expression values.}

\item{\code{control.normalized.visbound}}{control normalized data is thresholded in order to perform better visualization.}

\item{\code{control.normalized.visbound.noiseRemoved}}{noise is removed from control normalized and thresholded data.}

\item{\code{large.scale.cnv.events}}{large scale CNV events identified by CaSpER}

\item{\code{segments}}{CNV segments identified  by CaSpER}

\item{\code{cytoband}}{cytoband information downloaded from UCSC hg19: http://hgdownload.cse.ucsc.edu/goldenpath/hg19/database/cytoBand.txt.gz hg38:http://hgdownload.cse.ucsc.edu/goldenpath/hg38/database/cytoBand.txt.gz}

\item{\code{annotation}}{positions of each gene along each chromosome in the genome}

\item{\code{annotation.filt}}{lowly expressed genes are filtered from gene annotation data.frame}

\item{\code{control.sample.ids}}{vector containing the  reference (normal) cell (sample) names}

\item{\code{project.name}}{project name}

\item{\code{genomeVersion}}{genomeVersion: hg19 or hg38}

\item{\code{hmmparam}}{initial hmm parameters estimated from data}

\item{\code{plotorder}}{cell (sample) ordering for heatmap plots}

\item{\code{vis.bound}}{threshold for control normalized data for better visualization}

\item{\code{noise.thr}}{noise threshold for better visualization}

\item{\code{loh.name.mapping}}{containing the cell (sample) name and the matching baf signal sample name}

\item{\code{sequencing.type}}{sequencing type: bulk or single-cell}

\item{\code{cnv.scale}}{maximum expression scale}

\item{\code{loh.scale}}{maximum baf scale}

\item{\code{loh.shift.thr}}{baf shift threshold estimated from baf signal using gaussian mixture models}

\item{\code{window.length}}{window length used for median filtering}

\item{\code{length.iterations}}{increase in window length at each scale iteration}
}}

