% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CASPER_APP.R
\name{CASPER_APP}
\alias{CASPER_APP}
\title{Shiny App Function}
\usage{
CASPER_APP()
}
\description{
There are no arguments for this function, users will simply run the code CASPER_APP
and it will run the code and open the Shiny App. Other functions from the caspeR
package are used in this app. See details on GitHub for instructions on how to use
the Shiny App.
}
