% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{CenterSmooth}
\alias{CenterSmooth}
\title{CenterSmooth()}
\usage{
CenterSmooth(object)
}
\arguments{
\item{object}{casper object}
}
\value{
object
}
\description{
Cell centric expression centering is performed. For each cell (or sample), we compute the mid-point of the expression level then we subtract the mid-point expression from the expression levels of all the genes for the corresponding cell
}
