% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{ControlNormalize}
\alias{ControlNormalize}
\title{ControlNormalize()}
\usage{
ControlNormalize(object, vis.bound, noise.thr)
}
\arguments{
\item{object}{casper object}
}
\value{
object
}
\description{
The control normalization is performed by subtracting reference expression values from the tumor expression values.
}
