% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation.R
\name{extractEvents}
\alias{extractEvents}
\title{extractEvents()}
\usage{
extractEvents(segments, cytoband, type)
}
\arguments{
\item{cytoband}{cytoband information downloaded from UCSC hg19: http://hgdownload.cse.ucsc.edu/goldenpath/hg19/database/cytoBand.txt.gz hg38:http://hgdownload.cse.ucsc.edu/goldenpath/hg38/database/cytoBand.txt.gz}

\item{type}{event type amp (amplification) or del (deletion)/}

\item{object}{casper object}
}
\value{
combined large scale events in data.frame
}
\description{
formats large scale events as a matrix. Rows represent  samples (cells) whereas columns represent chromosome arms (1: amplification, 0: neutral, -1: deletion) 
helper function for generateLargeScaleEvents()
}
