% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation.R
\name{extractLargeScaleEvents}
\alias{extractLargeScaleEvents}
\title{extractLargeScaleEvents()}
\usage{
extractLargeScaleEvents(final.objects, thr = 0.5)
}
\arguments{
\item{final.objects}{casper object}

\item{thr}{gamma threshold determining the least number of scales required to support}
}
\value{
final large scale event summary reported as a matrix
}
\description{
generates coherent set of large scale CNV events using the pairwise comparison of all scales from BAF and expression signals
}
