% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{extractMUAndCooccurence}
\alias{extractMUAndCooccurence}
\title{extractMUAndCooccurence()}
\usage{
extractMUAndCooccurence(finalChrMat, loh, loh.name.mapping)
}
\arguments{
\item{finalChrMat}{large scale event matrix generated using CaSpER}

\item{loh}{original baf signal}

\item{loh.name.mapping}{contains the cell (sample) name and the matching baf signal sample name}
}
\value{
list of mutually exclusive and co-occurent events
}
\description{
calculates significant mutually exclusive and co-occurent events
}
