% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation.R
\name{extractSegmentSummary}
\alias{extractSegmentSummary}
\title{extractSegmentSummary()}
\usage{
extractSegmentSummary(final.objects)
}
\arguments{
\item{final.objects}{list of casper object}
}
\value{
list of loss and gain segments identified  in all scales
}
\description{
generates coherent set of CNV segments using the pairwise comparison of all scales from BAF and expression signals
}
