% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{gene.matrix}
\alias{gene.matrix}
\title{gene.matrix()}
\usage{
gene.matrix(segment, all.genes, all.samples, genes.ann)
}
\arguments{
\item{segment}{CNV segments}

\item{all.genes}{gene names}

\item{all.samples}{samp names}

\item{genes.ann}{gene symbols within each segments}
}
\value{
matrix of gene level CNV events
}
\description{
Gene level CNV events represented as matrix where rows represent samples and columns represent samples
}
