% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{generateAnnotation}
\alias{generateAnnotation}
\title{generateAnnotation()}
\usage{
generateAnnotation(id_type = "ensembl_gene_id", genes, ishg19,
  centromere)
}
\arguments{
\item{id_type}{gene list identifier, ensembl_gene_id or hgnc_symbol}

\item{genes}{list of genes}

\item{ishg19}{boolean values determining the genome version}

\item{centromere}{centromer regions}
}
\value{
list of mutually exclusive and co-occurent events
}
\description{
retrieves gene chromosomal locations from biomart
}
