% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{generateEnrichmentSummary}
\alias{generateEnrichmentSummary}
\title{generateEnrichmentSummary()}
\usage{
generateEnrichmentSummary(results)
}
\arguments{
\item{results}{output of getDiffExprGenes() function}
}
\value{
significantly enriched GO Terms
}
\description{
generate GO Term enrichment summary
}
