% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation.R
\name{generateLargeScaleEvents}
\alias{generateLargeScaleEvents}
\title{generateLargeScaleEvents()}
\usage{
generateLargeScaleEvents(object)
}
\arguments{
\item{object}{casper object}
}
\value{
object
}
\description{
generates large scale CNV events
}
