% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation.R
\name{generateParam}
\alias{generateParam}
\title{generateParam()}
\usage{
generateParam(object, cnv.scale = 3)
}
\arguments{
\item{object}{casper object}

\item{cnv.scale}{expression.scale for the expression signal}
}
\value{
object
}
\description{
Initial HMM parameters estimated from the data.
}
