% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_1_table_info_2020.R
\name{get_1_table_info_2020}
\alias{get_1_table_info_2020}
\title{Builds the scrapped table from the 2020 CASPER reports}
\usage{
get_1_table_info_2020(
  text,
  Facility_ID,
  CCN,
  Facility_Name,
  Report_Period,
  Comparision_Group,
  Run_Date,
  City_State,
  Version_Number
)
}
\arguments{
\item{Facility_ID}{Facility's ID Number in a string format, ex: "LC0011535", see
\code{get_facility_id()}}

\item{CCN}{CMS's certification number in a string format, ex: "L002.03", see
\code{get_CNN()}}

\item{Facility_Name}{Facility's name in string format, ex: "John's Fruit Stand", see
\code{get_facility_name()}}

\item{Report_Period}{The report period put in string format, month/day/year, ex: "04
/01/2023 - 09/30/2023", see \code{get_report_period_2020()}}

\item{Comparision_Group}{The comparison group put in string format, month/day/year,
ex: "02/02/2023 - 07/31/2023", see \code{get_comparison_group_2020()}}

\item{Run_Date}{The date that the report was compiled on put in string format, month
/day/year, ex: "10/06/2023", see \code{get_run_date()}}

\item{City_State}{This is the city and state that the facility is located in put in
string format, ex: "Richland, WA", see \code{get_city_and_state()}}

\item{Version_Number}{The report version number of the CASPER report in string
format, ex: "1.01", see \code{get_version_number()}}

\item{details}{the easiest way to use this function is to use the other functions
mentioned above to get the numbers from the pdf and assign variable names to
their outputs}

\item{object}{requires character strings, the raw_text from the CASPER pdf}
}
\value{
returns the all the information from the data table on the CASPER reports as
    a dataframe in R
}
\description{
This function is for pdf reports from 2020 that only have one table in the pdf
(Jan 2020 reports have one table but Feb 2020 reports have two). If the report being
scrapped has two tables see \code{get_two_tables_2020()}. The arguments for this
function have the same arguments as \code{get_table_info_old()} other than the
Report_Period and the Comparision_Group. User should also be aware that all NA's
are represented by the word "null".
}
\examples{
#make sure you have these packages installed and loaded
pacman::p_load(tidyverse, devtools, caspeR, pdftools, roxygen2,zoo, stringr, glue)

#there are two ways of using this function, putting the strings in by hand or naming
them as demonstrated in table1 and table2

raw_text <- pdf_text("R/caspeR_example.pdf")

table1 <- get_1_table_info_2020(raw_text, Facility_ID = "LC0011535", CCN = "L002.03"
, Facility_Name = "John's Fruit Stand", Report_Period = "04/01/2023 - 09/30/2023",
Comparision_Group = "02/02/2023 - 07/31/2023", Run_Date = "10/6/2023", City_State =
"Richland, WA",Version_Number = "1.01")

View(table1)

fac_ID <- get_facility_id(raw_text)
ccn <- get_CCN(raw_text)
fac_name <- get_facility_name(raw_text)
report_pd <- get_report_period_2020(raw_text)
comp_group <- get_comparison_group_2020(raw_text)
rundate <- get_run_date(raw_text)
citystate <- get_city_and_state(raw_text)
vnumber <- get_version_number(raw_text)

table2 <- get_1_table_info_2020(raw_text, Facility_ID = fac_ID,
CCN = ccn, Facility_Name = fac_name, Report_Period = report_pd,
Comparision_Group = comp_group, Run_Date = rundate, Version_Number = vnumber,
City_State = citystate)

View(table2)

}
