% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_report_identifiers_2020.R
\name{get_comparison_group_2020}
\alias{get_comparison_group_2020}
\title{Pdf scrapes comparison group from the 2020 CASPER reports}
\usage{
get_comparison_group_2020(text)
}
\arguments{
\item{object}{requires a character string (text) from the scrapped pdf}

\item{details}{may want to save as an object with the same name as shown below as these functions are built to go into building a table (see \code{\link{get_1_table_info_2020}})}
}
\value{
returns the comparison group from the pdf
}
\description{
The 2020 CASPER reports are slightly different from the 2019 reports (the date format is
different), so this function specifically scrapes 2020 CASPER reports
}
\examples{
    raw_text <- pdf_text("yourpdf.pdf")

    Comparison_Group <- get_comparison_group_2020(raw_text)

}
