% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_report_identifiers_2019.R
\name{get_facility_name}
\alias{get_facility_name}
\title{Pulls the Facility Name from the 2019-2020 CASPER reports}
\usage{
get_facility_name(text)
}
\arguments{
\item{object}{requires a character string (text) from the scrapped pdf}

\item{details}{may want to save as an object with the same name as shown in the example
below as these functions are needed to build a table (see \code{get_table_info_old})}
}
\value{
returns the Facility Name from the CASPER reports
}
\description{
This function can be used to pull the Facility Name off the pdfs, which is required for
building the table (see \code{get_table_info_old}).
}
\examples{
\dontrun{
raw_text <- pdf_text("yourpdf.pdf")

facility_name <- get_facility_name(raw_text)
}
}
