% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_report_identifiers_2019.R
\name{get_report_period}
\alias{get_report_period}
\title{Pulls the Report Period ONLY from the 2019 CASPER reports}
\usage{
get_report_period(text)
}
\arguments{
\item{object}{requires a character string (text) from the scrapped pdf}

\item{details}{may want to save as an object with the same name as shown in the example
below as these functions are needed to build a table (see \code{get_table_info_old})}
}
\value{
returns the Report Period from the 2019 CASPER reports
}
\description{
This function can be used to pull the Report Period off the pdfs, which is required for
building the table (see \code{get_table_info_old}). If it is a 2020 CASPER report
\code{get_report_period_2020} should be used (the pdf date format changed from 2019 to
2020).
}
\examples{
\dontrun{
raw_text <- pdf_text("yourpdf.pdf")

report_pd <- get_report_period(raw_text)
}
}
