% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_report_identifiers_2020.R
\name{get_report_period_2020}
\alias{get_report_period_2020}
\title{Pdf scrapes report period from the 2020 CASPER reports}
\usage{
get_report_period_2020(text)
}
\arguments{
\item{object}{requires a character string (text) from the scrapped pdf}

\item{details}{may want to save as an object as shown in the example below in
preparation for building a table}
}
\value{
returns the report period date range from the pdf as a string
}
\description{
The 2020 CASPER reports are slightly different from the 2019 reports (the date format
is different), so this function specifically scrapes report period from the 2020
CASPER reports which is required for building a table
(see \code{get_table_info_old()}, \code{get_1_table_info_2020()}, or
\code{get_2_tables_info_2020()} for further usage). If it is a 2019 CASPER report
\code{get_report_period()} should be used instead.
}
\examples{
#make sure you have these packages installed and loaded
pacman::p_load(tidyverse, devtools, caspeR, pdftools, roxygen2,zoo, stringr, glue)

raw_text <- pdf_text("R/caspeR_example.pdf")

report_pd <- get_report_period_2020(raw_text)

}
