% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{getDiffExprGenes}
\alias{getDiffExprGenes}
\title{getDiffExprGenes()}
\usage{
getDiffExprGenes(final.objects, sampleName, chrs, event.type)
}
\arguments{
\item{final.objects}{list of objects}

\item{sampleName}{sample name}

\item{chrs}{selected chromosomes}

\item{event.type}{cnv event type}
}
\value{
differentially expressed genes
}
\description{
get differentially expressed genes between samples having selected specified CNV events
}
