% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{goEnrichmentBP}
\alias{goEnrichmentBP}
\title{goEnrichmentBP()}
\usage{
goEnrichmentBP(genes, ontology, universe = character(0), pvalue = 0.05,
  annotation = "org.Hs.eg.db", conditionalSearch = TRUE, genes2)
}
\arguments{
\item{genes}{list of genes}

\item{ontology}{ontology  (BP, CC or MF)}

\item{universe}{universe of genes}

\item{pvalue}{pvalue cutoff}

\item{annotation}{ontology annotation default:org.Hs.eg.db}
}
\value{
significantly enriched GO Terms
}
\description{
GO Term enrichment
}
