% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loh.R
\name{lohCallMedianFilterByChr}
\alias{lohCallMedianFilterByChr}
\title{readBAFExtractOutput()}
\usage{
lohCallMedianFilterByChr(object, loh.scale, n = 50,
  scale.iteration = 50)
}
\arguments{
\item{path}{path for the folder that contains BAFExtract output files}
}
\value{
baf signal in data.frame format
}
\description{
Reads BAFExtract output files
}
