% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation.R
\name{mergeScalesAndGenerateFinalEventSummary}
\alias{mergeScalesAndGenerateFinalEventSummary}
\title{mergeScalesAndGenerateFinalEventSummary()}
\usage{
mergeScalesAndGenerateFinalEventSummary(final.objects)
}
\arguments{
\item{final.objects}{list of casper objects}
}
\value{
list of objects
}
\description{
helper function for extractLargeScaleEvents()
}
