% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{PerformMedianFilter}
\alias{PerformMedianFilter}
\title{Old R based implementation of ScaleData. Scales and centers the data}
\usage{
PerformMedianFilter(object, window.length = 50, length.iterations = 50)
}
\arguments{
\item{object}{casper object}

\item{genes.use}{Vector of gene names to scale/center. Default is all genes in object@data.}

\item{data.use}{Can optionally pass a matrix of data to scale, default is object@data[genes.use,]}

\item{do.scale}{Whether to scale the data.}

\item{do.center}{Whether to center the data.}

\item{scale.max}{Max value to accept for scaled data. The default is 10. Setting this can help
reduce the effects of genes that are only expressed in a very small number of cells.}
}
\value{
Returns a casper object with object@scale.data updated with scaled and/or centered data.
}
\description{
Old R based implementation of ScaleData. Scales and centers the data
}
\examples{
\dontrun{
pbmc_small <- ScaleDataR(object = pbmc_small)
}

}
