% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation.R
\name{PerformSegmentationWithHMM}
\alias{PerformSegmentationWithHMM}
\title{PerformSegmentationWithHMM()}
\usage{
PerformSegmentationWithHMM(object, cnv.scale, removeCentromere = T,
  cytoband)
}
\arguments{
\item{object}{casper object}

\item{cnv.scale}{expression signal scale number}

\item{removeCentromere}{boolean values determining if centromere regions should be removed from the analysis}

\item{cytoband}{cytoband information downloaded from UCSC hg19: http://hgdownload.cse.ucsc.edu/goldenpath/hg19/database/cytoBand.txt.gz hg38:http://hgdownload.cse.ucsc.edu/goldenpath/hg38/database/cytoBand.txt.gz}
}
\value{
object
}
\description{
HMM segmentation applied for each scale of expression signal
}
