% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotBAFAllSamples}
\alias{plotBAFAllSamples}
\title{plotBAFAllSamples()}
\usage{
plotBAFAllSamples(loh, fileName)
}
\arguments{
\item{loh}{baf signal, user can either give smoothed baf signal or original baf signal as an input.}

\item{fileName}{fileName of the putput image}
}
\description{
Visualization of BAF shift signal for all samples together
}
