% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotBAFInSeperatePages}
\alias{plotBAFInSeperatePages}
\title{plotBAFInSeperatePages()}
\usage{
plotBAFInSeperatePages(loh, folderName)
}
\arguments{
\item{loh}{baf signal, user can either give smoothed baf signal or original baf signal as an input.}

\item{folderName}{folder name for the output images}
}
\value{
object
}
\description{
Visualization of BAF deviation for each sample in separate pages
}
