% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotBAFOneSample}
\alias{plotBAFOneSample}
\title{plotBAFOneSample()}
\usage{
plotBAFOneSample(object, fileName)
}
\arguments{
\item{object}{casper object}

\item{fileName}{fileName of the output image}
}
\description{
Visualization of BAF shift signal in different scales for one sample
}
