% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotGEAllSamples}
\alias{plotGEAllSamples}
\title{plotGEAllSamples()}
\usage{
plotGEAllSamples(object, fileName = fileName, cnv.scale)
}
\arguments{
\item{object}{casper object}

\item{fileName}{fileName of the putput image}

\item{cnv.scale}{expression.scale for the expression signal}
}
\description{
plot gene expression signal for each sample seperately
}
