% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotGEAndBAFOneSample}
\alias{plotGEAndBAFOneSample}
\title{plotGEAndBAFOneSample()}
\usage{
plotGEAndBAFOneSample(object, cnv.scale, loh.scale, sample, n = 50,
  scale.iteration = 50)
}
\arguments{
\item{object}{casper object}

\item{cnv.scale}{expression.scale for the expression signal}

\item{sample}{sample name}

\item{n}{window length used for median filtering}

\item{length.iterations}{increase in window length at each scale iteration}
}
\description{
Gene expression and BAF signal for one sample in one plot
}
