% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotGEAndGT}
\alias{plotGEAndGT}
\title{plotGEAndGT()}
\usage{
plotGEAndGT(chrMat, genoMat, fileName)
}
\arguments{
\item{chrMat}{large scale events identified from CaSpER represented as matrix. Rows indicates samples (cells) whereas columns indicates chromosome arms}

\item{genoMat}{large scale events identified from genotyping array represented as matrix. Rows indicates samples (cells) whereas columns indicates chromosome arms}

\item{fileName}{fileName of the putput image}
}
\description{
Heatmap plot for large scale event calls identified by CaSpER and genotyping array.
}
