% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{plotHeatmap()}
\usage{
plotHeatmap(object, fileName, cnv.scale = 3, cluster_cols = F,
  cluster_rows = T, show_rownames = T, only_soi = T)
}
\arguments{
\item{object}{casper object}

\item{fileName}{fileName of the putput image}

\item{cnv.scale}{expression.scale for the expression signal}

\item{cluster_cols}{boolean values determining if columns should be clustered}

\item{cluster_rows}{boolean values determining if rows should be clustered}

\item{show_rownames}{boolean values determining if rownames should be plotted}

\item{only_soi}{boolean values determining if only samples of interest without control samples should be plotted}
}
\description{
Visualization of the genomewide gene expression signal plot at different smoothing scales
}
