% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotLargeScaleEvent}
\alias{plotLargeScaleEvent}
\title{plotLargeScaleEvent()}
\usage{
plotLargeScaleEvent(object, fileName)
}
\arguments{
\item{object}{casper object}

\item{fileName}{fileName of the output image}
}
\description{
Visualization of the  large-scale CNV events among all the samples/cells
}
