% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotLargeScaleEvent2}
\alias{plotLargeScaleEvent2}
\title{plotLargeScaleEvent2()}
\usage{
plotLargeScaleEvent2(chrMat, fileName)
}
\arguments{
\item{chrMat}{large scale events identified from CaSpER represented as matrix. Rows indicates samples (cells) whereas columns indicates chromosome arms}

\item{fileName}{fileName of the output image}
}
\description{
Visualization of the  large-scale CNV events among all the samples/cells
}
