% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotMUAndCooccurence}
\alias{plotMUAndCooccurence}
\title{plotMUAndCooccurence()}
\usage{
plotMUAndCooccurence(results)
}
\arguments{
\item{results}{output of extractMUAndCooccurence() function}
}
\description{
Visualization of mutually exclusive and co-occuring events
}
