% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotSCellCNVTree}
\alias{plotSCellCNVTree}
\title{plotSCellCNVTree()}
\usage{
plotSCellCNVTree(finalChrMat, sampleName,
  path = "C:\\\\Users\\\\aharmanci\\\\Downloads\\\\phylip-3.695\\\\phylip-3.695\\\\exe",
  fileName)
}
\arguments{
\item{finalChrMat}{large scale events identified from CaSpER represented as matrix. Rows indicates samples (cells) whereas columns indicates chromosome arms}

\item{sampleName}{sample name}

\item{path}{path to the executable containing fitch. If path = NULL, the R will search several commonly used directories for the correct executable file. More information about installing PHYLIP can be found on the PHYLIP webpage: http://evolution.genetics.washington.edu/phylip.html.}
}
\description{
Pyhlogenetic tree-based clustering and visualization of the cells based on the CNV events from single cell RNA-seq Data.
}
