% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotSingleCellLargeScaleEventHeatmap}
\alias{plotSingleCellLargeScaleEventHeatmap}
\title{plotSingleCellLargeScaleEventHeatmap()}
\usage{
plotSingleCellLargeScaleEventHeatmap(finalChrMat, sampleName, chrs)
}
\arguments{
\item{finalChrMat}{large scale events identified from CaSpER represented as matrix. Rows indicates samples (cells) whereas columns indicates chromosome arms}

\item{sampleName}{sample name}

\item{chrs}{chromosome names}
}
\value{
object
}
\description{
Visualization of large scale event summary for selected samples and chromosomes
}
