\name{procBam}
\alias{procBam}
\title{
  Process BAM object
}
\description{
  Process paired-end data stored in BAM object generated by scanBam. 
  Outputs RangedData object.
  }
\usage{
procBam(bam, stranded=FALSE, seed)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{bam}{
    BAM object generated by \code{scanBam}
  }
  \item{stranded}{
    Set to \code{TRUE} to indicate that the RNA-seq experiment preserved
    the strand information.
  }
  \item{seed}{
    Seed for random number generator
  }
}
\details{
One alignment is chosen at random from those multihits with same start position for both reads but different insertions/deletions patterns. 
}
\value{
  An object of class \code{procBam} containing 
  reads with both ends
  correctly aligned and split according to the corresponding CIGAR.
  Unique identifiers by fragment are stored.}
\author{
  Camille Stephan-Otto Attolini
}
\seealso{
\code{scanBam} from package \code{Rsamtools}
}
\examples{
##See example in calcExp
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ manip }

