% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{ProcessData}
\alias{ProcessData}
\title{ProcessData()}
\usage{
ProcessData(object)
}
\arguments{
\item{object}{casper object}
}
\value{
object
}
\description{
Processing expression signal. Step 1. Recursively iterative median filtering  Step 2. Center Normalization Step 3. Control Normalization
}
