% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loh.R
\name{readBAFExtractOutput}
\alias{readBAFExtractOutput}
\title{readBAFExtractOutput()}
\usage{
readBAFExtractOutput(path, sequencing.type = "bulk")
}
\arguments{
\item{path}{path for the folder that contains BAFExtract output files}
}
\value{
baf signal in data.frame format
}
\description{
Reads BAFExtract output files
}
