% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casper.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,casper-method}
\title{show method for casper}
\usage{
\S4method{show}{casper}(object)
}
\arguments{
\item{object}{A casper object}
}
\description{
show method for casper
}
