% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation.R
\name{splitByOverlap}
\alias{splitByOverlap}
\title{splitByOverlap()}
\usage{
splitByOverlap(query, subject, column = "ENTREZID", ...)
}
\description{
helper function for segment summary. Acknowledgements to https://support.bioconductor.org/p/67118/
}
