% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_report_identifiers_2019.R
\name{get_report_period}
\alias{get_report_period}
\title{Pulls the report period ONLY for the 2019 CASPER reports}
\usage{
get_report_period(text)
}
\arguments{
\item{object}{requires a character string (text) from the scrapped pdf}

\item{details}{may want to save as an object as shown in the example below in
preparation for building a table}
}
\value{
returns the report period date range from the 2019 CASPER reports as a
   string
}
\description{
This function can be used to pull the report period off the 2019 pdfs, which is
required for building a table (see \code{get_table_info_old()},
\code{get_1_table_info_2020()}, or \code{get_2_tables_info_2020()} for further usage
).If it is a 2020 CASPER report \code{get_report_period_2020()} should be used (the
pdf date format changed from 2019 to 2020).
}
\examples{
#make sure you have these packages installed and loaded
pacman::p_load(tidyverse, devtools, caspeR, pdftools, roxygen2,zoo, stringr, glue)

raw_text <- pdf_text("R/caspeR_example.pdf")

report_pd <- get_report_period(raw_text)

}
