% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_report_identifiers_2019.R
\name{get_run_date}
\alias{get_run_date}
\title{Pulls the Run Date from the 2019-2020 CASPER reports}
\usage{
get_run_date(text)
}
\arguments{
\item{object}{requires a character string (text) from the scrapped pdf}

\item{details}{may want to save as an object with the same name as shown in the example
below as these functions are needed to build a table (see \code{\link{get_table_info_old}
})}
}
\value{
returns the Run Date from the CASPER reports
}
\description{
This function can be used to pull the Run Date off the pdfs, which is required for
building the table (see \code{get_table_info_old}).
}
