% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table_info_old.R
\name{get_table_info_old}
\alias{get_table_info_old}
\title{Builds the scrapped table from the CASPER reports prior to 2020}
\usage{
get_table_info_old(
  text,
  Facility_ID,
  CCN,
  Facility_Name,
  Report_Period,
  Comparision_Group,
  Run_Date,
  City_State,
  Version_Number
)
}
\arguments{
\item{object}{requires character strings, the raw_text from the CASPER pdf for the
following: Facility_ID, CCN, Facility_Name, Report_Period, Comparision_Group,
Run_Date, City_State, and the Version_Number}

\item{details}{the easiest way to use this function is to use the other functions
mentioned above to get the numbers from the pdf and assign variable names to their
outputs}
}
\value{
returns the all the information from the data table on the CASPER reports prior to
2020 as a dataframe in R
}
\description{
The arguments for this function are the outputs obtained from using the following
functions: \code{get_facility_id}, \code{get_CNN}, \code{get_facility_name}, \code{get_rep
ort_period}, \code{get_comparison_group}, \code{get_run_date}, \code{get_city_and_state
},\code{get_version_number}.
User should also be aware that all NA's are represented by the word "null"
}
\examples{
   raw_text <- pdf_text("yourpdf.pdf")
   get_table_info_old(raw_text, "MS0004", "123456", "Name of Facility", "09/01/2019 - 02
   /31/2020", "10/01/2019 - 03/31/2020", "11/03/2019", "Rexburg, ID","3.02")

   raw_text <- pdf_text("yourpdf.pdf")
   Facility_ID <- get_facility_id(raw_text)
   CNN <- get_CNN(raw_text)

   get_table_info_old(raw_text, Facility_ID, CCN, Facility_Name, Report_Period, Comparisio
   n_Group, Run_Date, City_State, Version_Number)

}
