% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{PerformMedianFilter}
\alias{PerformMedianFilter}
\title{PerformMedianFilter()}
\usage{
PerformMedianFilter(object, window.length = 50, length.iterations = 50)
}
\arguments{
\item{object}{casper object}

\item{window.length}{window length used for median filtering}

\item{length.iterations}{increase in window length at each scale iteration}
}
\value{
object
}
\description{
Recusive iterative median filtering is applied to whole genome
}
